Exptree Implementation is Complete. Used Double instead of float throughout the project.

+++ Parsers/SQLCreateParser.c
@@ -4,7 +4,7 @@
 < table-name > ::=  identifier
 < columns > ::= <column> | <column> , <columns>
 < column> ::= identifer <dtype> |   identifer <dtype> <primary key | not null >
-< dtype> ::= varchar(<digits>) | int | float
+< dtype> ::= varchar(<digits>) | int | double
 < digits> := <digit> | <digit><digits>
 < digit> ::= 0|1|2|3|4|5|6|7|8|9
 */
